refactor(shared): move hex color helper into shared module
- move normalizeHexColor utility from logic/ to shared/

- update theme and preference imports to shared/hexColor

- keep validation behavior unchanged while aligning folder ownership
