+++ widget-src/shared/hexColor.ts
